refactor: move WaveformVisualizer to a separate component, delete unused component
