+++ packages/audiodocs/src/components/InteractivePlayground/index.tsx
@@ -8,85 +8,6 @@ import Reset from "@site/static/img/reset.svg";
 import ResetDark from "@site/static/img/reset-dark.svg";
 import AnimableIcon, { Animation } from "@site/src/components/AnimableIcon";
 
-interface RangeSliderProps {
-  label: string;
-  value: number;
-  min: number;
-  max: number;
-  step?: number;
-  unit?: string;
-  onChange: (v: number) => void;
-}
-export const RangeSlider: FC<RangeSliderProps> = ({
-  label,
-  value,
-  min,
-  max,
-  step = 1,
-  unit = "",
-  onChange,
-}) => (
-  <div className={styles.rangeRow}>
-    <label className={styles.rangeLabel}>
-      {label}: {Number(value).toFixed(2)} {unit}
-    </label>
-    <input
-      type="range"
-      className={styles.slider}
-      value={value}
-      min={min}
-      max={max}
-      step={step}
-      onChange={(e) => onChange(parseFloat(e.target.value))}
-    />
-  </div>
-);
-
-interface SelectProps {
-  label: string;
-  value: string;
-  options: string[];
-  onChange: (v: string) => void;
-}
-export const Select: FC<SelectProps> = ({
-  label,
-  value,
-  options,
-  onChange,
-}) => (
-  <div className={styles.selectRow}>
-    <label className={styles.selectLabel}>{label}</label>
-    <select
-      className={styles.select}
-      value={value}
-      onChange={(e) => onChange(e.target.value)}
-    >
-      {options.map((opt) => (
-        <option key={opt} value={opt}>
-          {opt}
-        </option>
-      ))}
-    </select>
-  </div>
-);
-
-interface SwitchProps {
-  label: string;
-  value: boolean;
-  onChange: (v: boolean) => void;
-}
-export const CustomSwitch: FC<SwitchProps> = ({ label, value, onChange }) => (
-  <label className={styles.switchRow}>
-    <input
-      type="checkbox"
-      className={styles.checkbox}
-      checked={value}
-      onChange={(e) => onChange(e.target.checked)}
-    />
-    <span className={styles.switchLabel}>{label}</span>
-  </label>
-);
-
 interface PlaygroundHookResult {
   example: FC<any>;
   props: Record<string, any>;
